refactor: extract metadata utils
