+++ packages/common/src/utils/media.ts
@@ -1,8 +1,6 @@
 import type { Playlist, PlaylistItem } from '../../types/playlist';
 import { CONTENT_TYPE } from '../constants';
 
-import { formatDuration, formatVideoSchedule } from './formatting';
-
 type RequiredProperties<T, P extends keyof T> = T & Required<Pick<T, P>>;
 
 type DeprecatedPlaylistItem = {
@@ -51,39 +49,3 @@ export const getLegacySeriesPlaylistIdFromEpisodeTags = (item: PlaylistItem | un
 
 export const isLiveChannel = (item: PlaylistItem): item is RequiredProperties<PlaylistItem, 'contentType' | 'liveChannelsId'> =>
   item.contentType?.toLowerCase() === CONTENT_TYPE.liveChannel && !!item.liveChannelsId;
-
-export const createVideoMetadata = (media: PlaylistItem, episodesLabel?: string) => {
-  const metaData = [];
-  const duration = formatDuration(media.duration);
-
-  if (media.pubdate) metaData.push(String(new Date(media.pubdate * 1000).getFullYear()));
-  if (!episodesLabel && duration) metaData.push(duration);
-  if (episodesLabel) metaData.push(episodesLabel);
-  if (media.genre) metaData.push(media.genre);
-  if (media.rating) metaData.push(media.rating);
-
-  return metaData;
-};
-
-export const createPlaylistMetadata = (playlist: Playlist, episodesLabel?: string) => {
-  const metaData = [];
-
-  if (episodesLabel) metaData.push(episodesLabel);
-  if (playlist.genre) metaData.push(playlist.genre as string);
-  if (playlist.rating) metaData.push(playlist.rating as string);
-
-  return metaData;
-};
-
-export const createLiveEventMetadata = (media: PlaylistItem, locale: string) => {
-  const metaData = [];
-  const scheduled = formatVideoSchedule(locale, media.scheduledStart, media.scheduledEnd);
-  const duration = formatDuration(media.duration);
-
-  if (scheduled) metaData.push(scheduled);
-  if (duration) metaData.push(duration);
-  if (media.genre) metaData.push(media.genre);
-  if (media.rating) metaData.push(media.rating);
-
-  return metaData;
-};
